# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Public API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get list of articles",
                "operationId": "b7a11d15ce1e37223a10814d716316ec",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Text to search in articles",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255,
                            "example": "The Art of Letting Go"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArticleResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – Bearer token missing or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Add a new article",
                "operationId": "9ffef4b4a0e32f17945d68a234598a27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuthKeysResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Article not saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/articles/{slug}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Retrieve a single article",
                "operationId": "293de0571f0175dc769961d78b036256",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the article",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "the-art-of-letting-go"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – Bearer token missing or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update an existing article",
                "description": "Updates the specified article using the given data",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "my-new-title"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Article not saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login useig keys",
                "operationId": "c903e13854d8c70027b4b84a0c4aeadf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AppKeyAuthRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authed user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuthKeysResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred."
                    }
                },
                "type": "object"
            },
            "AppKeyAuthRequest": {
                "title": "Login with public key and secret key Request",
                "description": "Request parameters to identify user has access",
                "properties": {
                    "app_key": {
                        "description": "Generated Public Key",
                        "type": "string",
                        "example": "3ebd11afbf17bf7780a76a0288262cc05dd8848d"
                    },
                    "app_secret": {
                        "description": "Generated Secret Key",
                        "type": "string",
                        "example": "445d314af62141f846b2fa547521e9ff70c36b11"
                    }
                },
                "type": "object"
            },
            "SearchArticlesRequest": {
                "title": "List Articles Request",
                "description": "Request parameters for listing articles",
                "properties": {
                    "search": {
                        "description": "The article search",
                        "type": "string",
                        "example": "The Art of Letting Go "
                    }
                },
                "type": "object"
            },
            "StoreArticleRequest": {
                "title": "New Article Data Request",
                "description": "New articles data",
                "properties": {
                    "title": {
                        "description": "The article title",
                        "type": "string",
                        "example": "My New Title"
                    },
                    "subtitle": {
                        "description": "The article subtitle",
                        "type": "string",
                        "example": "My Subtitle"
                    },
                    "body": {
                        "description": "The article body",
                        "type": "string",
                        "example": "Some body text"
                    },
                    "slug": {
                        "type": "string",
                        "example": "my-new-title-2"
                    }
                },
                "type": "object"
            },
            "UpdateArticleRequest": {
                "title": "Update Article Data Request",
                "description": "Update articles data",
                "properties": {
                    "title": {
                        "description": "The article title",
                        "type": "string",
                        "example": "My New Title"
                    },
                    "subtitle": {
                        "description": "The article subtitle",
                        "type": "string",
                        "example": "My Subtitle"
                    },
                    "body": {
                        "description": "The article body",
                        "type": "string",
                        "example": "Some body text"
                    },
                    "slug": {
                        "type": "string",
                        "example": "my-new-title"
                    }
                },
                "type": "object"
            },
            "ArticleResource": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "The Art of Letting Go: Why Detachment Brings Peace"
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Understanding the freedom in releasing control and expectations"
                    },
                    "body": {
                        "type": "string",
                        "example": "In a world that constantly encourages us to hustle, chase, and control outcomes, the idea of letting go can feel like giving up. But true peace and growth often begin when we detach from what we can't control. By surrendering the need for a fixed result, you free yourself from unnecessary anxiety. You allow life to unfold naturally, and often, in better ways than you imagined. Practicing detachment is not weakness; it's strength rooted in trust and emotional maturity."
                    },
                    "slug": {
                        "type": "string",
                        "example": "the-art-of-letting-go"
                    },
                    "destination": {
                        "$ref": "#/components/schemas/DestinationResource"
                    }
                },
                "type": "object"
            },
            "AuthKeysResource": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "token": {
                        "type": "string",
                        "example": "19|z3fbgx1XiDS8CSjGFUJ2edj5Ef1rcnQDGe8P99yt0b21740b"
                    },
                    "name": {
                        "type": "string",
                        "example": "xhovana"
                    }
                },
                "type": "object"
            },
            "DestinationResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Destination 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}